added ability to deposit/withdraw with floating point, added try/catch clauses to input parsing

+++ src/main/java/rocks/zipcode/atm/CashMachine.java
@@ -29,7 +29,7 @@ public class CashMachine {
         );
     }
 
-    public void deposit(int amount) {
+    public void deposit(float amount) {
         if (accountData != null) {
             tryCall(
                     () -> bank.deposit(accountData, amount),
